+++ ui/store/modules/system/actions.js
@@ -73,7 +73,7 @@ export default {
                   wsMessage.stimulation_protocols_running.includes(true)) ||
                 wsMessage.system_status === SYSTEM_STATUS.OFFLINE_STATE;
 
-              commit(
+              dispatch(
                 "stimulation/setStimStatus",
                 stimPlayState ? STIM_STATUS.STIM_ACTIVE : STIM_STATUS.READY
               );
@@ -136,14 +136,17 @@ export default {
     const result = TextValidationPlateBarcode.validate(newValue, type);
     const isValid = result == "";
 
-    // stop all running processes if either barcode changes regardless of validity
-    if (this.state.stimulation.stimPlayState) {
-      await this.dispatch("stimulation/stopStimulation");
-      commit("setBarcodeWarning", true);
-    }
-    // require new stim configuration check if either new barcode changes
-    if (isValid && state.barcodes[type].value !== newValue) {
-      this.commit("stimulation/setStimStatus", STIM_STATUS.CONFIG_CHECK_NEEDED);
+    // if this is the first barcode scanned, can skip these
+    if (state.barcodes[type].value) {
+      // stop all running processes if either barcode changes regardless of validity
+      if (this.state.stimulation.stimPlayState) {
+        await this.dispatch("stimulation/stopStimulation");
+        commit("setBarcodeWarning", true);
+      }
+      // require new stim configuration check if either new barcode changes
+      if (isValid && state.barcodes[type].value !== newValue) {
+        await this.dispatch("stimulation/setStimStatus", STIM_STATUS.CONFIG_CHECK_NEEDED);
+      }
     }
 
     commit("setBarcode", { type, newValue, isValid });